% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data from inside.airbnb.com}
\usage{
clean_data(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
x dataframe
}
\description{
It filters for verified hosts only and set the minimum requirement of reviews to 4.
This function is designed for Danish data only as it transforms the price-related variables in a Danish Kroner format.
Lastly, it filters for a price per night between the following range: [0-7400] which approximately displays [0-1000 Euro]
}
\examples{
dataset_abc <- data.frame("host_identity_verified" = c(TRUE,FALSE,TRUE), "number_of_reviews" = c(1,4,6),
"cleaning_fee" = c("$20","$25", "$22"),
"price" = c("$100","$200", "$300"), "extra_people"= c("$3", "$6", "$2"),
security_deposit= c("$1000","$2002", "$3003"), stringsAsFactors=FALSE)
clean_data(dataset_abc)

}
