% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names_nh.R
\name{clean_names_nh}
\alias{clean_names_nh}
\title{Cleaning and correction of mispellings}
\usage{
clean_names_nh(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
dataframe with its new variable
}
\description{
Detect the presence of misspellings and add the correct form in a new variable called
Neighbourhood.
}
\examples{
dataset_abc <- data.frame(
"neighbourhood_cleansed" = c("Amager st","Amager Vest", "Vanlse", "Nrrebro$"),
"first_review" = c("2018-10-17", "2018-01-18", "2016-10-05", "2016"),
stringsAsFactors=FALSE)
clean_names_nh(dataset_abc)
}
