% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance from two points or two vectors}
\usage{
distance(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{Longitude of first data set}

\item{lat1}{Latitude of first data set}

\item{long2}{Longitude of second data set}

\item{lat2}{Latitude of second data set}
}
\value{
a vector of distances
}
\description{
Distance from two points or two vectors
}
\examples{
distance(c(12.12345,14.14244),
 c(14.12345, 15.2355),
 50.2345,
 11.223415)
}
