% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrainfo.R
\name{extrainfo}
\alias{extrainfo}
\title{Extra Informations:
It calculates the price per person per night and the duration the listings is available.}
\usage{
extrainfo(x)
}
\arguments{
\item{x}{dataframe (after the use of \code{\link{clean_data}})}
}
\value{
dataframe with two new variables
}
\description{
Extra Informations:
It calculates the price per person per night and the duration the listings is available.
}
\examples{
 dataset_abc <- data.frame(
  "price_dkk" = c(200, 300, 400),
 "accommodates"= c(2,5,20),
 "last_scraped" = c("2018-10-18","2018-10-18", "2018-10-18"),
 "first_review" = c("2018-10-17", "2018-01-18", "2016-10-05"),
 stringsAsFactors=FALSE)
extrainfo(dataset_abc)
}
