% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_metro.R
\name{geo_metro}
\alias{geo_metro}
\title{Scraping Geo-Coordinates from Wikipedia}
\usage{
geo_metro(x)
}
\arguments{
\item{x}{URL}
}
\value{
geo coordinates
}
\description{
This functions aims at wikipedia tables that include links (hrefs) to other wikipedia sites
where the coordinates are located.
}
\examples{
metroURL <- "https://en.wikipedia.org/wiki/List_of_Copenhagen_Metro_stations"
geo_metro(metroURL)
}
