% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Review Index}
\usage{
index(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
dataframe with a new variable "index"
}
\description{
This function creates an addivtive index of six numeric variables which all relate to ratings concerning an Airbnb listing.
See the details in order to see which variables are included in the index.
}
\details{
\itemize{
\item  review_scores_accuracy
\item  review_scores_cleanliness
\item  review_scores_checkin
\item  review_scores_communication
\item  review_scores_location
\item  review_scores_value
}
}
\examples{


}
