% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data2}
\alias{load_data2}
\title{Read in csv files}
\usage{
load_data2(infile)
}
\arguments{
\item{infile}{= the file name}
}
\value{
a dataframe
}
\description{
Read in csv files
}
\examples{

dataset_abc <- data.frame("SN" = 1:2, "Age" = c(21,15), "Name" = c("John","Dora"), stringsAsFactors=FALSE)
dataset_csv <- write.table(dataset_abc, "filename.csv", sep = ",", row.names = F)
load_data2("filename.csv")

}
