% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cph.R
\name{map_cph}
\alias{map_cph}
\title{Map of Copenhagen with some attributes}
\usage{
map_cph(infile, infile2)
}
\arguments{
\item{infile}{dataframe of metro coordinates and names of metro stations}

\item{infile2}{dataframe of listing coordinates}
}
\value{
a leaflet map
}
\description{
Map of Copenhagen with some attributes
}
