% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cph_nb.R
\name{map_cph_nb}
\alias{map_cph_nb}
\title{Mapping Neighbourhoods Copenhagen}
\usage{
map_cph_nb(infile, infile2)
}
\arguments{
\item{infile}{geo data (polygon)}

\item{infile2}{data frame including longitutde and latitude of listings}
}
\value{
a leaflet map including polygons of neighbourhoods
}
\description{
Mapping Neighbourhoods Copenhagen
}
