% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot Density Price per Night}
\usage{
plot_density(infile)
}
\arguments{
\item{x}{dataframe}
}
\value{
a density plot
}
\description{
It takes the log of the price and plots a density plot using ggplot2. It requires the variables
price_dkk and rtype
}
\examples{
df <- data.frame(
"price_dkk" = c(200,500,1000),
"rtype" = c("Apartment", "Private Room", "Apartment"),
stringsAsFactors=FALSE)
plot_density(df)


}
