% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance.R
\name{plot_distance}
\alias{plot_distance}
\title{Plot Distance to next metro station}
\usage{
plot_distance(infile)
}
\arguments{
\item{infile}{}
}
\value{
a ggplot
}
\description{
Plotting the distance to the next metro station depending on its price (log)
}
