% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_c.R
\name{plot_distance_c}
\alias{plot_distance_c}
\title{Plot Distance to city center}
\usage{
plot_distance_c(infile)
}
\arguments{
\item{infile}{}
}
\value{
a ggplot
}
\description{
Plotting the distance to the next metro station depending on its price (log)
Requires the variables price_dkk and dist_centrum
}
