% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_price_apart_full.R
\name{plot_price_apart_full}
\alias{plot_price_apart_full}
\title{Night price distribution of entire apartments AirBnB CPH}
\usage{
plot_price_apart_full(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a ggplot (histogram)
}
\description{
Night price distribution of entire apartments AirBnB CPH
}
\examples{
dataset_abc <- data.frame(
"room_type" = c("Entire home/apt","Entire home/apt","Private room"),
"price" = c("$100","$1500", "$300"), stringsAsFactors=FALSE)
plot_price_apart_full(dataset_abc)
}
