% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_cph.R
\name{regression_cph}
\alias{regression_cph}
\title{Estimate models with robust standard errors}
\usage{
regression_cph(Data)
}
\arguments{
\item{Data}{data frame}
}
\value{
4 regressions and 2 latex ouput
}
\description{
Estimate models with robust standard errors
}
