% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_table.R
\name{regression_table}
\alias{regression_table}
\title{Regression Table}
\usage{
regression_table(infile, infile2, infile3, infile4, infile5)
}
\arguments{
\item{infile}{Regression Model}

\item{infile2}{Regression Model}

\item{infile3}{Regression Model}

\item{infile4}{Regression Model}

\item{infile5}{Regression Model}
}
\value{
stargazer table output html
}
\description{
Regression Table
}
