% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Recoding inside dataset}
\usage{
tidy_data(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
x dataframe
}
\description{
This function creates dummy variables of four variables (see the example)
}
\examples{
dataset_abc <- data.frame(
  "room_type" = c("Entire home/apt","Private room","Private room"),
 "host_is_superhost" = c("t","f", "t"),
 "instant_bookable" = c("t","t", "f"),
 "cancellation_policy" = c("strict_14_with_grace_period", "moderate", "flexible" )
 stringsAsFactors=FALSE)
tidy_data(dataset_abc)

}
